Far too many test code for one method
This is a nice exercise. More and more, I feel like we don’t need to
unit test everything. It is far easier to inspect the code and see if
it is correct. On the other hand, unit testing everything *really*
makes sure your code is correct, and stays correct, while explicitly
stating the coder’s intention. Still, I wrote that method with pen in
paper in a minute or so on paper. Implementing it with the amount of
tests necessary increased that time fifteenfold.